x/authentication/types/message_authenticate_user_test.go
package types

import (
	"testing"

	"identity/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgAuthenticateUser_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgAuthenticateUser
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgAuthenticateUser{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgAuthenticateUser{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
